pyproject.toml
[project]
name = "chatbox"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "argon2-cffi>=25.1.0",
    "email-validator>=2.3.0",
    "fastapi>=0.128.8",
    "python-jose>=3.3.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.11.0",
    "sqlalchemy>=2.0.48",
    "uvicorn>=0.39.0",
    "python-multipart>=0.0.20",
]
